+++ pyintelliclima/__init__.py
@@ -2,22 +2,28 @@ from .api import (
     IntelliClimaAPI,
     IntelliClimaAPIError,
     IntelliClimaAuthError,
+    IntelliClimaEcocomfort3API,
     IntelliClimaEcocomfortAPI,
 )
 from .intelliclima_types import (
     IntelliClimaC800,
     IntelliClimaDevices,
     IntelliClimaECO,
+    IntelliClimaECO3,
     IntelliClimaLoginBody,
+    IntelliClimaVMCBase,
 )
 
 __all__ = (
     "IntelliClimaEcocomfortAPI",
+    "IntelliClimaEcocomfort3API",
     "IntelliClimaAPI",
     "IntelliClimaAPIError",
     "IntelliClimaAuthError",
     "IntelliClimaDevices",
     "IntelliClimaC800",
     "IntelliClimaECO",
+    "IntelliClimaECO3",
+    "IntelliClimaVMCBase",
     "IntelliClimaLoginBody",
 )